requirements.txt
ultralytics==8.4.11
opencv-python==4.13.0.90
torch==2.10.0
torchvision==0.25.0
numpy==2.4.2
paho-mqtt
firebase-admin
psutil